feat: Add configuration for running Client scripts
This commit adds a new configuration in the `.fleet/run.json` file for running the Client.py script. It also adds a new configuration in the file for running the Client.py script with a different name. The configurations include the working directory and file path for the scripts.

Refactor Client.py

This commit updates the Client.py script by removing unnecessary comments and code. It replaces the input function for entering the admin password with a dialog box. It also adds functionality to handle commands entered as messages. If the admin user enters "/kick" followed by a username, the script sends a "KICK" message to the server to kick the specified user. If the admin user enters "/ban" followed by a username, the script sends a "BAN" message to the server to ban the specified user. Non-admin users will receive a message that commands can only be executed by an admin.

Refactor Server.py

This commit updates the Server.py script to handle the "KICK" and "BAN" messages received from the client. If a message starts with "KICK" or "BAN", the script extracts the username and performs the corresponding action. If a user is kicked, a message is broadcasted to inform other clients. If a user is banned, their username is added to the bans.txt file and a message is printed.

docs: Add new test script

This commit adds a new test.py script to the repository. The script assigns a sample string to a variable and demonstrates extracting a substring using indexing.